video_helpers/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
decorator==4.4.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
moviepy==1.0.3
numpy==2.0.0
openai==1.35.7
packaging==24.1
Pillow==9.5.0
proglog==0.1.10
pydantic==2.7.4
pydantic_core==2.18.4
python-env==1.0.0
replicate==0.27.0
requests==2.32.3
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
